requirements.txt
attrs>=21.4.0
cachetools>=4.2.2
certifi>=2021.5.30
charset-normalizer>=2.0.6
click>=7.1.2
cloudevents>=1.2.0
cryptography>=35.0.0
deprecation>=2.1.0
flask>=1.1.2
functions-framework>=2.2.1
google-auth>=2.2.0
google-cloud-logging>=2.6.0
google-cloud-storage>=1.42.3
google-cloud-monitoring>=2.6.0
google-cloud-secret-manager>=2.7.2
gunicorn>=20.0.4
idna>=3.2
itsdangerous>=1.1.0
jinja2>=2.11.2
markupsafe>=1.1.1
packaging>=21.0
paramiko>=2.8.0
pathtools>=0.1.2
pip>=21.1.3
pyasn1-modules>=0.2.8
pyasn1>=0.4.8
pyparsing>=2.4.7
requests>=2.26.0
rsa>=4.7.2
setuptools>=58.1.0
stringcase>=1.2.0
urllib3>=1.26.7
watchdog>=1.0.2
werkzeug>=1.0.1